improving queries and helper methods, mosly for optimization and code duplication

+++ entity/movie.go
@@ -4,7 +4,7 @@ type Movie struct {
 	Id uint
 	Title string
 	ReleaseYear int
-	Duration float64
+	Duration int
 
 	Actors []Actor
 	Genres []Genre
